services/function-registration/requirements.txt
fastapi
pydantic
psycopg2
kafka-python
uvicorn
prometheus_client
structlog